requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.9.3
asgiref==3.5.2
astunparse==1.6.3
attrs==22.1.0
black==22.10.0
boto3==1.26.14
botocore==1.29.14
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==2.2.0
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
dataclasses==0.6
Django==4.1.3
fire==0.4.0
flatbuffers==22.11.23
fonttools==4.38.0
future==0.18.2
fvcore==0.1.5.post20221122
gast==0.4.0
gif2numpy==1.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
hydra-core==1.2.0
idna==3.4
imageio==2.22.4
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==5.0.0
iopath==0.1.10
jmespath==1.0.1
jsonschema==4.17.1
kaitaistruct==0.10
keras==2.11.0
kiwisolver==1.4.4
labelme2coco==0.2.1
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mypy-extensions==0.4.3
networkx==2.8.8
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pathspec==0.10.2
Pillow==9.3.0
pixellib==0.7.1
platformdirs==2.5.4
portalocker==2.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.5
pycocotools==2.0.6
pydot==1.4.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.19.2
python-dateutil==2.8.2
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
sahi==0.11.4
scikit-image==0.19.3
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminaltables==3.1.10
tifffile==2022.10.10
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
xmljson==0.2.1
yacs==0.1.8
zipp==3.10.0
